feat: enhance website model and API for transaction history 💰✨
- Added depositHistory and withdrawHistory arrays to track transactions 📈📉
- Updated POST and PUT routes to handle new transaction data 🛠️
- Improved AddWebsiteForm to manage transaction types (Deposit/Withdraw) 🔄
- Enhanced UI with better spacing and button sizes for a more polished look 🎨
