+++ .vscode/settings.json
@@ -4,7 +4,7 @@
     ],
     "github.copilot.chat.commitMessageGeneration.instructions": [
       {
-        "text": "Be exteremely detailed with the changes and give in pointers. and add lots of emojis!",
+        "text": "Be short but precise in pointers. give priority to logical changes first then design based changes and add lots and lots of emojis with some sarcastic commets!!",
       }
     ]
 }
