+++ app/api/websites/[name]/route.js
@@ -23,16 +23,23 @@ export async function PUT(request, context) {
     await connection();
     // Await the params from the context.
     const { name } = await context.params;
-    const { account_number, url } = await request.json();
+    const { account_number, url, website_name, current_balance, type } =
+      await request.json();
     // Uncomment and modify the update operation as needed:
+    console.log("type", type);
     await Website.updateOne(
       { website_name: name },
       {
         $set: {
+          website_name: website_name,
           account_number,
+          current_balance,
           url,
+          depositHistory: type ? [current_balance] : [],
+          withdrawHistory: type ? [] : [current_balance],
         },
-      }
+      },
+      { upsert: true }
     );
     return NextResponse.json({
       Message: `Website account updated successfully`,
