+++ components/AddWebsiteForm.jsx
@@ -5,18 +5,22 @@ import React, { useEffect, useState } from "react";
 import { FaUniversity, FaCode, FaMoneyBillWave } from "react-icons/fa";
 import { toast } from "react-toastify";
 import FullScreenLoader from "./FullScreenLoader";
+import { DropdownMenu } from "./DropdownMenu";
 
 const AddWebsiteForm = ({ setShowAddWebsiteForm, fetchData, editData }) => {
   const [websiteName, setWebsiteName] = useState("");
   const [url, setUrl] = useState("");
+  const [initialWebsiteName, setInitialWebsiteName] = useState("");
+  const [initialBalance, setInitialBalance] = useState("");
   const [currentBalance, setCurrentBalance] = useState("");
   const [loading, setLoading] = useState(false);
+  const [transactionType, setTransactionType] = useState("Deposit");
 
   const handleEdit = async (data) => {
     setLoading(true);
     try {
       const response = await axios.put(
-        `/api/websites/${data.website_name}`,
+        `/api/websites/${initialWebsiteName}`,
         data
       );
       toast.success(response?.data?.message);
@@ -54,6 +58,7 @@ const AddWebsiteForm = ({ setShowAddWebsiteForm, fetchData, editData }) => {
         website_name: websiteName,
         url: url,
         current_balance: currentBalance,
+        type: transactionType === "Deposit" ? true : false,
       });
       setShowAddWebsiteForm(false);
       fetchData();
@@ -108,6 +113,8 @@ const AddWebsiteForm = ({ setShowAddWebsiteForm, fetchData, editData }) => {
       setWebsiteName(editData.website_name);
       setUrl(editData.url);
       setCurrentBalance(editData.current_balance);
+      setInitialBalance(editData.current_balance);
+      setInitialWebsiteName(editData?.website_name);
     }
   }, [editData]);
 
@@ -127,8 +134,6 @@ const AddWebsiteForm = ({ setShowAddWebsiteForm, fetchData, editData }) => {
               className="appearance-none bg-transparent border-none w-full text-gray-700 py-1 px-2 leading-tight focus:outline-none"
               type="text"
               placeholder="Website Name"
-              disabled={editData}
-              style={{ cursor: editData ? "not-allowed" : "default" }}
               aria-label="Website Name"
               value={websiteName}
               onChange={(e) => setWebsiteName(e.target.value)}
@@ -153,14 +158,29 @@ const AddWebsiteForm = ({ setShowAddWebsiteForm, fetchData, editData }) => {
               className="appearance-none bg-transparent border-none w-full text-gray-700 py-1 px-2 leading-tight focus:outline-none"
               type="number"
               step="0.1"
-              disabled={editData}
-              style={{ cursor: editData ? "not-allowed" : "default" }}
               placeholder="Current Balance"
               aria-label="Current Balance"
               value={currentBalance}
-              onChange={(e) => setCurrentBalance(e.target.value)}
+              onChange={(e) => {
+                setCurrentBalance(e.target.value);
+                if (editData) {
+                  setTransactionType(
+                    Number(e.target.value) > initialBalance
+                      ? "Deposit"
+                      : "Withdraw"
+                  );
+                }
+              }}
             />
           </div>
+
+          <DropdownMenu
+            label=""
+            options={["Deposit", "Withdraw"]}
+            value={transactionType}
+            onChange={setTransactionType}
+            isDisabled={true}
+          />
 
           <div className="flex items-center justify-center">
             <button
